+++ tests/test_package.py
@@ -2,4 +2,4 @@ import teremok
 
 
 def test_version() -> None:
-    assert teremok.__version__ == "0.2.0"
+    assert teremok.__version__ == "0.3.0"